requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.163.0
spacy==3.8.4
numpy==2.2.3
torch==2.6.0